moving util tests into new format, breaking out fixture for createdomain
